+++ app/controllers/concerns/api_responders.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module ApiResponders
+  extend ActiveSupport::Concern
+
+  private
+
+    def respond_with_error(message, status = :unprocessable_entity, context = {})
+      is_exception = message.kind_of?(StandardError)
+      error_message = is_exception ? message.record&.errors_to_sentence : message
+      render status: status, json: { error: error_message }.merge(context)
+    end
+
+    def respond_with_success(message, status = :ok, context = {})
+      render status: status, json: { notice: message }.merge(context)
+    end
+
+    def respond_with_json(json = {}, status = :ok)
+      render status: status, json: json
+    end
+end
